requirements.txt
requests
django==2.2.6
pytest
pytest-pythonpath==0.7.3
pytest-django
djangorestframework
djangorestframework-simplejwt
python-slugify==5.0.2
django-filter==2.4.0
